+++ arex-schedule-web-api/src/main/java/com/arextest/schedule/model/plan/PostSendRequest.java
@@ -10,8 +10,9 @@ import lombok.Data;
 @Data
 public class PostSendRequest {
   @NotNull
-  private String caseId;
+  private String planId;
   @NotNull
+  private String caseId;
   private String replayId;
   @NotNull
   private Integer sendStatusType;
