+++ arex-schedule-web-api/src/main/java/com/arextest/schedule/model/plan/preSendRequest.java
@@ -10,6 +10,10 @@ import lombok.Data;
 @Data
 public class preSendRequest {
   @NotNull
+  private String planId;
+  @NotNull
+  private String caseId;
+  @NotNull
   private int replayPlanType;
   @NotNull
   private String recordId;
